% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos.R
\docType{data}
\name{duos}
\alias{duos}
\title{List of duos}
\format{An object of class \code{list} of length 20.}
\usage{
duos
}
\description{
List containing all 20 duo palettes, defined by
hexadecimal values.
}
\keyword{datasets}
