% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quads.R
\docType{data}
\name{quads}
\alias{quads}
\title{List of quads}
\format{An object of class \code{list} of length 20.}
\usage{
quads
}
\description{
List containing all 20 quad palettes, defined by
hexadecimal values.
}
\keyword{datasets}
