% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{sanzo.demo.all}
\alias{sanzo.demo.all}
\title{Print demo base R plots for all 60 sanzo palettes to illustrate
their use and display hexadecimal values.}
\usage{
sanzo.demo.all()
}
\description{
Print demo base R plots for all 60 sanzo palettes to illustrate
their use and display hexadecimal values.
}
\examples{
sanzo.demo.all()

}
