% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{sanzo.demo2}
\alias{sanzo.demo2}
\title{Print demo base R plots for all 20 duo palettes to illustrate
their use and display hexadecimal values.}
\usage{
sanzo.demo2()
}
\description{
Print demo base R plots for all 20 duo palettes to illustrate
their use and display hexadecimal values.
}
\examples{
sanzo.demo2()

}
