% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sanzo.info.all}
\alias{sanzo.info.all}
\title{See metadata for all palettes}
\usage{
sanzo.info.all()
}
\value{
Return data frame containing long-form names, short-form IDs,
hex values, and URLs for all sanzo palettes.
}
\description{
Create data frame containing long-form names, short-form IDs,
hexadecimal values, and links to Dain M. Blodorn Kim's
https://sanzo-wada.dmbk.io for all sanzo palettes.
}
\examples{
info_df <- sanzo.info.all()
print(sanzo.info.all())

}
