% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sanzo.info2}
\alias{sanzo.info2}
\title{Create data frame containing long-form names, short-form IDs,
hexadecimal values, and links to Dain M. Blodorn Kim's 
https://sanzo-wada.dmbk.io for all duo palettes.}
\usage{
sanzo.info2()
}
\value{
Return data frame containing long-form names, short-form IDs,
hex values, and URLs for all duos.
}
\description{
Create data frame containing long-form names, short-form IDs,
hexadecimal values, and links to Dain M. Blodorn Kim's 
https://sanzo-wada.dmbk.io for all duo palettes.
}
\examples{
duo_info_df <- sanzo.info2()
print(sanzo.info2())
knitr::kable(sanzo.info2())

}
