% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sanzo.info3}
\alias{sanzo.info3}
\title{See metadata for all trio palettes}
\usage{
sanzo.info3()
}
\value{
Return data frame containing long-form names, short-form IDs,
hex values, and URLs for all trios.
}
\description{
Create data frame containing long-form names, short-form IDs,
hexadecimal values, and links to Dain M. Blodorn Kim's
https://sanzo-wada.dmbk.io for all trio palettes.
}
\examples{
trio_info_df <- sanzo.info3()
print(sanzo.info3())
knitr::kable(sanzo.info3())

}
