% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sanzo.info4}
\alias{sanzo.info4}
\title{See metadata for all quad palettes}
\usage{
sanzo.info4()
}
\value{
Return data frame containing long-form names, short-form IDs,
hex values, and URLs for all quads
}
\description{
Create data frame containing long-form names, short-form IDs,
hexadecimal values, and links to Dain M. Blodorn Kim's
https://sanzo-wada.dmbk.io for all quad palettes.
}
\examples{
quad_info_df <- sanzo.info4()
print(sanzo.info4())
knitr::kable(sanzo.info4())

}
