% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios.R
\docType{data}
\name{trios}
\alias{trios}
\title{List containing all 20 trio palettes, defined by
hexadecimal values.}
\format{An object of class \code{list} of length 20.}
\usage{
trios
}
\description{
List containing all 20 trio palettes, defined by
hexadecimal values.
}
\keyword{datasets}
