% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{sanzo.demo3}
\alias{sanzo.demo3}
\title{Print demo base R plots for all 20 trio palettes to illustrate
their use and display hexadecimal values.}
\usage{
sanzo.demo3()
}
\description{
Print demo base R plots for all 20 trio palettes to illustrate
their use and display hexadecimal values.
}
\examples{
sanzo.demo3()

}
