% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos.R
\name{sanzo.duo}
\alias{sanzo.duo}
\title{Generate two-colored palettes}
\usage{
sanzo.duo(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "c006".}
}
\value{
A vector of hexademicals of length 2.
}
\description{
Generate two-colored palettes.
}
\examples{
# Assign palette to a name.
my_palette <- sanzo.duo("c229")

# Concatenate two duos for a custom quad.
c033 <- sanzo.duo("c033")
c095 <- sanzo.duo("c095")
custom_quad <- c(c033, c095)

# Use with base R.
plot(iris$Sepal.Width, iris$Sepal.Length, col = sanzo.duo("c085"))

# Use with ggplot2.
ggplot(df, aes(fill = Concentration, y = Cq, x = Target)) +
   geom_bar(stat = "identity") +
   theme_few() +
   scale_fill_manual(values = sanzo.duo("c103"))

# Use as gradient.
ggplot(faithfuld, aes(waiting, eruptions)) +
   geom_raster(aes(fill = density), interpolate = TRUE) +
   theme_few() +
   scale_fill_gradientn(colors = sanzo.duo("c102"))

}
