% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quads.R
\name{sanzo.quad}
\alias{sanzo.quad}
\title{Generate four-colored palettes}
\usage{
sanzo.quad(paletteName)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "c263".}
}
\description{
Generate four-colored palettes.
}
\examples{
# Assign palette to a name.
my_palette <- sanzo.quad("c252")

# Concatenate two quads for a custom eight-colored palette.
c348 <- sanzo.quad("c348")
c341 <- sanzo.quad("c341")
custom_eight <- c(c348, c341)

# Use with base R.
plot(df$x, df$y, col = sanzo.quad("c341"))

# Use with ggplot2.
ggplot(df, aes(fill = Concentration, y = Cq, x = Target)) +
   geom_bar(stat = "identity") +
   theme_few() +
   scale_fill_manual(values = sanzo.quad("c296"))

# Use as gradient.
ggplot(faithfuld, aes(waiting, eruptions)) +
   geom_raster(aes(fill = density), interpolate = TRUE) +
   theme_few() +
   scale_fill_gradientn(colors = sanzo.quad("c302"))

}
