% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios.R
\name{sanzo.trio}
\alias{sanzo.trio}
\title{Generate three-colored palettes.}
\usage{
sanzo.trio(paletteName)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "c121".}
}
\description{
Generate three-colored palettes.
}
\examples{
# Save palette with name.
my_palette <- sanzo.trio("c223")

# Concatenate two trios for a custom six-colored palette.
c207 <- sanzo.trio("c207")
c226 <- sanzo.trio("c226")
custom_six <- c(c207, c226)

# Use with base R.
plot(df$x, df$y, col = sanzo.trio("c343"))

# Use with ggplot2.
ggplot(df, aes(fill = Concentration, y = Cq, x = Target)) + 
   geom_bar(stat = "identity") +
   theme_few() +
   scale_fill_manual(values = sanzo.trio("c172"))

# Use as gradient.
ggplot(faithfuld, aes(waiting, eruptions)) +
   geom_raster(aes(fill = density), interpolate = TRUE) +
   theme_few() +
   scale_fill_gradientn(colors = sanzo.trio("c240"))

}
